refactor: renaming it to the template {feature}Service.ts
